[ecore][ts] move EObjectInternal class to its own module and define constants
