+++ ecore/tsconfig.json
@@ -1,7 +1,7 @@
 {
   "compilerOptions": {
     "module": "commonjs",
-    "target": "es2020",
+    "target": "es2017",
     "sourceMap": true,
     "declaration": true,
     "outDir": "dist",
